refactor: create file for processing dirigeants pp
