+++ data_aggregation/dirigeants_pp.py
@@ -1,0 +1,50 @@
+import pandas as pd
+
+
+def preprocess_dirigeants_pp(query):
+    cols = [column[0] for column in query.description]
+    rep_chunk = pd.DataFrame.from_records(data=query.fetchall(), columns=cols)
+    rep_chunk.sort_values(
+        by=[
+            "siren",
+            "nom_patronymique",
+            "nom_usage",
+            "prenoms",
+            "datenaissance",
+            "villenaissance",
+            "paysnaissance",
+            "qualite",
+        ],
+        inplace=True,
+        ascending=[True, False, False, False, False, False, False, False],
+    )
+    rep_chunk.drop_duplicates(
+        subset=[
+            "siren",
+            "nom_patronymique",
+            "nom_usage",
+            "prenoms",
+            "datenaissance",
+            "villenaissance",
+            "paysnaissance",
+            "qualite",
+        ],
+        keep="first",
+        inplace=True,
+    )
+    rep_clean = (
+        rep_chunk.groupby(
+            by=[
+                "siren",
+                "nom_patronymique",
+                "nom_usage",
+                "prenoms",
+                "datenaissance",
+                "villenaissance",
+                "paysnaissance",
+            ]
+        )["qualite"]
+        .apply(lambda x: ", ".join(x))
+        .reset_index()
+    )
+    return rep_clean
